ElectionMaps/requirements.txt
beautifulsoup4==4.9.0
pywikibot==3.0.20200405
PyYAML==5.3.1
lxml=4.5.1
